Fixed image upload in add property function. Allowed for preview update in update property. Now when admin delete a user, said user's properties are deleted. Admin can't delete themselves anymore. Updated sql with admin login Added report to README

+++ model/deleteuser.php
@@ -5,12 +5,19 @@
 		//if user id is 1 ignore
 		if($_REQUEST['deluser'] !='1'){
 			include("../include/config.php");
+			if($_SESSION["id"] == $_REQUEST['deluser']){
 
-			$stmt = $db->prepare('DELETE FROM users WHERE id = :id') ;
-			$stmt->execute(array(':id' => $_REQUEST['deluser']));
+			}else{
+				$stmt = $db->prepare('DELETE FROM users WHERE id = :id') ;
+				$stmt->execute(array(':id' => $_REQUEST['deluser']));
 
-			header('Location: ../profile.php');
-			exit;
+				$stmt = $db->prepare('DELETE FROM properties WHERE user_id = :id') ;
+				$stmt->execute(array(':id' => $_REQUEST['deluser']));
+
+				header('Location: ../profile.php');
+				exit;
+			}
+			
 
 		}
 	}
